requirements.txt
google-cloud-storage
ruamel.yaml
